[REFACTOR] Folder structure redone for `tests`
