+++ tests/e2e/paths.js
@@ -4,8 +4,8 @@
 
 import path from 'path';
 
-const rayCharlesBaseFile = path.resolve('./tests/wikipedia_en_ray_charles_2015-06.zimaa');
-const gutenbergRoBaseFile = path.resolve('./tests/gutenberg_ro_all_2023-05.zim');
+const rayCharlesBaseFile = path.resolve('./tests/zims/legacy-ray-charles/wikipedia_en_ray_charles_2015-06.zimaa');
+const gutenbergRoBaseFile = path.resolve('./tests/zims/gutenberg-ro/gutenberg_ro_all_2023-05.zim');
 const downloadDir = path.resolve('./tests/');
 
 export default {
