+++ tests/e2e/runners/edge/ieMode.e2e.runner.js
@@ -1,7 +1,7 @@
 import { Builder } from 'selenium-webdriver';
 import { Options } from 'selenium-webdriver/ie.js';
-import legacyRayCharles from './legacy-ray_charles.e2e.spec.js';
-import gutenbergRo from './gutenberg_ro.e2e.spec.js';
+import legacyRayCharles from '../../specs/legacy-ray_charles.e2e.spec.js';
+import gutenbergRo from '../../specs/gutenberg_ro.e2e.spec.js';
 
 /* eslint-disable camelcase */
