+++ tests/e2e/runners/firefox/firefox.e2e.runner.js
@@ -1,8 +1,8 @@
 import { Builder } from 'selenium-webdriver';
 import firefox from 'selenium-webdriver/firefox.js';
-import legacyRayCharles from './legacy-ray_charles.e2e.spec.js';
-import gutenbergRo from './gutenberg_ro.e2e.spec.js';
-import paths from './paths.js';
+import legacyRayCharles from '../../specs/legacy-ray_charles.e2e.spec.js';
+import gutenbergRo from '../../specs/gutenberg_ro.e2e.spec.js';
+import paths from '../../paths.js';
 
 /* eslint-disable camelcase */
