+++ tests/e2e/runners/firefox/firefox56.bs.runner.js
@@ -1,26 +1,26 @@
 import { Builder } from 'selenium-webdriver';
-import legacyRayCharles from './legacy-ray_charles.e2e.spec.js';
+import legacyRayCharles from '../../specs/legacy-ray_charles.e2e.spec.js';
 
 /* eslint-disable camelcase */
 
 // Input capabilities
 const capabilities = {
     'bstack:options': {
-        os: 'OS X',
-        osVersion: 'Big Sur',
-        browserVersion: '14.1',
+        os: 'Windows',
+        osVersion: '10',
+        browserVersion: '56.0',
         projectName: 'BStack Project Name: Kiwix JS e2e tests',
-        buildName: 'BStack Build Name: Safari 14 on Big Sur',
+        buildName: 'BStack Build Name: Firefox Legacy 56',
         local: true,
         localIdentifier: process.env.BROWSERSTACK_LOCAL_IDENTIFIER,
         userName: process.env.BROWSERSTACK_USERNAME,
         accessKey: process.env.BROWSERSTACK_ACCESS_KEY,
-        seleniumVersion: '4.10.0'
+        seleniumVersion: '3.141.59'
     },
-    browserName: 'Safari'
+    browserName: 'Firefox'
 };
 
-async function loadSafariDriver () {
+async function loadFirefoxDriver () {
     const driver = await new Builder()
         // .forBrowser('edge')
         .usingServer('https://hub-cloud.browserstack.com/wd/hub')
@@ -29,10 +29,11 @@ async function loadSafariDriver () {
     return driver;
 };
 
-const driver_safari = await loadSafariDriver();
+const driver_fx = await loadFirefoxDriver();
 
 // Maximize the window so that full browser state is visible in the screenshots
-await driver_safari.manage().window().maximize();
-// Browserstack Safari does not support Service Workers
-console.log('Running tests in jQuery mode only for this browser version')
-legacyRayCharles.runTests(driver_safari, ['jquery']);
+await driver_fx.manage().window().maximize();
+
+// Run tests in jQuery mode only (this browser does not support Service Workers)
+console.log('Running tests in jQuery mode only for this browser version');
+legacyRayCharles.runTests(driver_fx, ['jquery']);
