+++ tests/e2e/specs/legacy-ray_charles.e2e.spec.js
@@ -23,7 +23,7 @@
 import { By, Key, WebDriver } from 'selenium-webdriver';
 // import firefox from 'selenium-webdriver/firefox.js';
 import assert from 'assert';
-import paths from './paths.js';
+import paths from '../paths.js';
 
 /* eslint-disable camelcase, one-var, prefer-const */
 /* global describe, it */
@@ -40,7 +40,7 @@ const port = process.env.BROWSERSTACK_LOCAL_IDENTIFIER ? '8099' : '8080';
 let rayCharlesBaseFile = paths.rayCharlesBaseFile;
 // For BrowserStack, we have to construct the file blops with XHR instead
 if (BROWSERSTACK) {
-    rayCharlesBaseFile = '/tests/wikipedia_en_ray_charles_2015-06.zimaa';
+    rayCharlesBaseFile = '/tests/e2e/zims/legacy-ray-charles/wikipedia_en_ray_charles_2015-06.zimaa';
 }
 let rayCharlesAllParts = '', rayCharlesFileArray = [];
 for (let i = 0; i < 15; i++) {
