+++ tests/unit/index.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta charset="utf-8">
-        <base href="../">
+        <base href="../../">
         <title>Kiwix-js Unit tests</title>
         
         <!--
@@ -30,8 +30,8 @@
         <script>
             QUnit.config.autostart = false;
         </script>
-        <script type="text/javascript" src="./tests/init.js"></script>
-        <script type="module" src="./tests/tests.js"></script>
+        <script type="text/javascript" src="./tests/unit/init.js"></script>
+        <script type="module" src="./tests/unit/tests.js"></script>
     </head>
     <body>
         <b>NOTE:</b> Firefox and Chrome do not allow access to some local filesystem files used in testing. So, if you're opening this through a file:// URL, you should instead go through a web server : either through a local one (http://localhost/...) or through a remote one (but you need SSL : https://webserver/...).<br/>
