+++ tests/zims/legacy-ray-charles/wikipedia_en_ray_charles_2015-06.zimaj
